2.1.0: -executor: Refactored. -server: Uninitialize modules AFTER everything has been destroyed. -video_format: use std::size_t for video_format_desc.size. git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@2450 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ accelerator/ogl/image/image_mixer.cpp
@@ -103,7 +103,7 @@ public:
 			auto buffer = spl::make_shared<const std::vector<uint8_t, tbb::cache_aligned_allocator<uint8_t>>>(format_desc.size, 0);
 			return async(launch::deferred, [=]
 			{
-				return core::const_array(buffer->data(), static_cast<std::size_t>(format_desc.size), true, buffer);
+				return core::const_array(buffer->data(), format_desc.size, true, buffer);
 			});
 		}
